🔄 refactor(activity-synthesis): extract source-agnostic synthesis to its own module
Moves ActivitySynthesis type and deriveSynthesisFromSignals function (plus
classifyThemes / sortThemes / looksLike* helpers) out of src/activity-summary.ts
into a new source-neutral module src/activity-synthesis.ts. Pure relocation —
no behavior change. activity-summary.ts now imports the synthesis function and
the two exported helpers (classifyThemes, sortThemes) from the new module.
Test import path updated in tests/activity-summary.test.ts.

Refs: UNC-136
